error handling dependency helper

+++ app/Helpers/DependencyHelper.php
@@ -11,10 +11,13 @@ namespace Vanderbilt\EpicParticipantUpdater\App\Helpers;
 
     private function getComposer()
     {
-        exec('php -r "copy(\'https://getcomposer.org/installer\', \'composer-setup.php\');"');
-        exec('php -r "if (hash_file(\'SHA384\', \'composer-setup.php\') === \'93b54496392c062774670ac18b134c3b3a95e5a5e5c8f1a9f115f203b75bf9a129d5daa8ba6a13e2cc8a1da0806388a8\') { echo \'Installer verified\'; } else { echo \'Installer corrupt\'; unlink(\'composer-setup.php\'); } echo PHP_EOL;"');
-        exec('php composer-setup.php');
-        exec('php -r "unlink(\'composer-setup.php\');"');
+        if(!file_exists('composer.phar'))
+        {
+            exec('php -r "copy(\'https://getcomposer.org/installer\', \'composer-setup.php\');"');
+            exec('php -r "if (hash_file(\'SHA384\', \'composer-setup.php\') === \'93b54496392c062774670ac18b134c3b3a95e5a5e5c8f1a9f115f203b75bf9a129d5daa8ba6a13e2cc8a1da0806388a8\') { echo \'Installer verified\'; } else { echo \'Installer corrupt\'; unlink(\'composer-setup.php\'); } echo PHP_EOL;"');
+            exec('php composer-setup.php');
+            exec('php -r "unlink(\'composer-setup.php\');"');
+        }
     }
 
     public function installDependencies()
